refactor(whispering): adopt the shared sign-in migration kit
Behavior-identical: openLocalSource, the prose (describe, the audio-files
note, the recordings error noun) stay app-side; the probe, copy-then-clear,
and dialog phases now come from @epicenter/app-shell/sign-in-migration. The
shared table iteration covers the same three tables the hardcoded list did.
